+++ R/utils-expr.R
@@ -578,6 +578,53 @@ translate <- function(
           }
           attr(out, "case_insensitive") <- case_insensitive
           return(out)
+        },
+
+        "%>%" = {
+          lhs <- expr[[2]]
+          rhs <- expr[[3]]
+          replace_dot <- function(expr, replacement) {
+            if (is.symbol(expr) && identical(expr, sym("."))) {
+              return(replacement)
+            } else if (is.call(expr)) {
+              as.call(lapply(expr, replace_dot, replacement = replacement))
+            } else {
+              return(expr)
+            }
+          }
+
+          has_dot <- any(vapply(
+            rhs,
+            function(x) is.symbol(x) && identical(x, sym(".")),
+            FUN.VALUE = logical(1)
+          ))
+
+          if (has_dot) {
+            new_rhs <- replace_dot(rhs, lhs)
+          } else {
+            ### I want to insert `lhs` as first arg so first I remove all args
+            ### in the original call (after saving them), and then I insert the
+            ### new args with `lhs` first.
+            existing_args <- call_args(rhs)
+            remove_args <- rep_named(names(existing_args), list(zap()))
+            new_args <- append(list(lhs), existing_args)
+            new_rhs <- rhs
+            if (length(remove_args) > 0) {
+              new_rhs <- call_modify(new_rhs, !!!remove_args)
+            }
+            new_rhs <- call_modify(new_rhs, !!!new_args)
+          }
+
+          out <- translate(
+            new_rhs,
+            .data = .data,
+            new_vars = new_vars,
+            env = env,
+            caller = caller,
+            call_is_function = call_is_function,
+            expr_uses_col = expr_uses_col
+          )
+          return(out)
         }
       )
 
@@ -675,6 +722,14 @@ translate <- function(
               "`tidypolars` doesn't know how to translate this function: `",
               fn_names$orig_name,
               "()`."
+            )
+          }
+          # Only suggest opening an issue for functions coming from other pkgs,
+          # not for custom functions.
+          if (!is.null(fn_names$pkg) || grepl("::", fn_names$orig_name)) {
+            msg <- c(
+              msg,
+              i = "You can ask for it to be translated here: <https://github.com/etiennebacher/tidypolars/issues>."
             )
           }
           abort(msg, call = env)
@@ -875,9 +930,15 @@ add_pkg_suffix <- function(name, known_ops, user_defined) {
 
   if (is.null(pkg) || pkg %in% c("base", "stats", "utils", "tools")) {
     name_to_eval <- paste0("pl_", fn)
-    pkg <- NULL
   } else {
     name_to_eval <- paste0("pl_", fn, "_", pkg)
+  }
+
+  if (grepl("::", name)) {
+    # Don't store the pkg name if it's explicitly specified.
+    # Don't do this too early because we need pkg when building the function to
+    # eval.
+    pkg <- NULL
   }
   list(orig_name = name, name_to_eval = name_to_eval, pkg = pkg)
 }