+++ tests/testthat/test-magrittr-pipe-lazy.R
@@ -1,0 +1,47 @@
+### [GENERATED AUTOMATICALLY] Update test-magrittr-pipe.R instead.
+
+Sys.setenv('TIDYPOLARS_TEST' = TRUE)
+
+test_that("%>% works in expression without '.'", {
+  test <- pl$LazyFrame(x = 1:3)
+  expect_equal_lazy(
+    test |> mutate(y = x %>% mean()),
+    data.frame(x = 1:3, y = 2)
+  )
+  expect_equal_lazy(
+    test |> mutate(y = x %>% mean(na.rm = TRUE)),
+    data.frame(x = 1:3, y = 2)
+  )
+})
+
+test_that("%>% works in expression with '.'", {
+  test <- pl$LazyFrame(x = 1:3)
+  expect_equal_lazy(
+    test |> mutate(y = x %>% mean(x = .)),
+    data.frame(x = 1:3, y = 2)
+  )
+})
+
+test_that("%>% works with summarize()", {
+  test <- pl$LazyFrame(x = 1:3)
+  expect_equal_lazy(
+    test |> summarize(y = x %>% mean(x = .)),
+    data.frame(y = 2)
+  )
+})
+
+test_that("chaining %>% works", {
+  test <- pl$LazyFrame(x = 1:3)
+  expect_equal_lazy(
+    test |> mutate(y = x %>% sqrt() %>% mean(na.rm = TRUE)),
+    data.frame(x = 1:3, y = 1.3820),
+    tolerance = 1e-4
+  )
+  expect_equal_lazy(
+    test |> mutate(y = x %>% sqrt(x = .) %>% mean(x = ., na.rm = TRUE)),
+    data.frame(x = 1:3, y = 1.3820),
+    tolerance = 1e-4
+  )
+})
+
+Sys.setenv('TIDYPOLARS_TEST' = FALSE)