+++ NEWS.md
@@ -73,6 +73,11 @@
 
 * Add support for `as.Date()` for character columns (#190).
 
+* Error messages due to untranslated functions now suggest opening an issue to
+  ask for their translation (#197).
+
+* Add support for `%>%` in expressions (#200).
+
 ## Bug fixes
 
 * Fix behavior of `mutate()` and `summarize()` when they don't contain any